Use portable format specifiers: Case %ld - plain - vim_snprintf.
Fix uses of plain "%ld" within vim_snprintf():
- Replace "%ld" with "%" PRId64.
- Cast corresponding argument to (int64_t).

+++ src/screen.c
@@ -7900,10 +7900,10 @@ static void win_redr_ruler(win_T *wp, int always)
      * Some sprintfs return the length, some return a pointer.
      * To avoid portability problems we use strlen() here.
      */
-    vim_snprintf((char *)buffer, RULER_BUF_LEN, "%ld,",
+    vim_snprintf((char *)buffer, RULER_BUF_LEN, "%" PRId64 ",",
         (wp->w_buffer->b_ml.ml_flags & ML_EMPTY)
-        ? 0L
-        : (long)(wp->w_cursor.lnum));
+        ? (int64_t)0L
+        : (int64_t)(wp->w_cursor.lnum));
     len = STRLEN(buffer);
     col_print(buffer + len, RULER_BUF_LEN - len,
         empty_line ? 0 : (int)wp->w_cursor.col + 1,
